+++ src/DbTests.DbGen/SQL/TempTableDefinitions/SpatialTypesTempTable.sql
@@ -1,0 +1,7 @@
+CREATE TABLE #SpatialTypesTempTable
+(
+  [Shape] GEOMETRY NOT NULL,
+  [NullableShape] GEOMETRY NULL,
+  [Location] GEOGRAPHY NOT NULL,
+  [NullableLocation] GEOGRAPHY NULL
+)