+++ src/TestDb/dbo/Stored Procedures/ProcWithSpatialTypesFromTvp.sql
@@ -1,0 +1,5 @@
+CREATE PROCEDURE [dbo].[ProcWithSpatialTypesFromTvp]
+  @params dbo.SpatialTypesTableType READONLY
+AS
+
+SELECT * FROM @params