+++ src/TestDb/dbo/Tables/TableWithSpatialTypes.sql
@@ -1,0 +1,8 @@
+CREATE TABLE [dbo].[TableWithSpatialTypes]
+(
+  [key] INT NOT NULL PRIMARY KEY,
+  [shape] GEOMETRY NOT NULL,
+  [nullableShape] GEOMETRY NULL,
+  [location] GEOGRAPHY NOT NULL,
+  [nullableLocation] GEOGRAPHY NULL
+)