+++ src/TestDb/dbo/User Defined Types/SpatialTypesTableType.sql
@@ -1,0 +1,7 @@
+CREATE TYPE [dbo].[SpatialTypesTableType] AS TABLE
+(
+  [shape] GEOMETRY NOT NULL,
+  [nullableShape] GEOMETRY NULL,
+  [location] GEOGRAPHY NOT NULL,
+  [nullableLocation] GEOGRAPHY NULL
+)